refactor: split external api namespaces
