+++ SgfDevs/Dev/EventSync/EventSyncImportFilter.cs
@@ -2,6 +2,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using SgfDevs.Dev.EventSync.Sessionize;
 
 namespace SgfDevs.Dev.EventSync;
