+++ SgfDevs/Dev/EventSync/MeetupEventMatcher.cs
@@ -3,6 +3,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
+using SgfDevs.Dev.EventSync.Meetup;
 
 namespace SgfDevs.Dev.EventSync;
