+++ SgfDevs/Dev/EventSync/Sessionize/SessionizeApiClient.cs
@@ -8,7 +8,7 @@ using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 
-namespace SgfDevs.Dev.EventSync;
+namespace SgfDevs.Dev.EventSync.Sessionize;
 
 public class SessionizeApiClient
 {
