+++ SgfDevs/Dev/EventSync/SessionizeSyncPlanner.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using SgfDevs.Dev.EventSync.Sessionize;
 
 namespace SgfDevs.Dev.EventSync;
